refactor criteria build
